+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- ``dtoolcore.generate_admin_metadata`` function raises
+  ``dtoolcore.DtoolCoreInvalidNameError`` if invalid name is provided
 - ``dtoolcore.utils.name_is_valid`` utility function for checking sanity of
   dataset names